backend/requirements.txt
fastapi
uvicorn
python-dotenv
redis
motor
groq
requests
pydantic[email]
pydantic-settings
scikit-learn
numpy
pandas
sse-starlette
bcrypt
passlib[bcrypt]
python-jose[cryptography]
python-multipart
uvloop; sys_platform != 'win32'
httpx
email-validator
